Make LogViewer test work regardless of time zone
The LogViewer test uses some date objects, which were in UTC.
When the component is rendered, the date is shown in local
time, which varies depending on the time zone.

Changing the tests so that the date objects are both created
and rendered in local time. This will ensure that the date is
rendered consistently across all time zones.